Implement simple Windows-1252 decoder
The encoding crate has many encodings which aren't selectable as features. It
takes quite a while to build. We also don't need encoding, only decoding,
because the messages DLL cannot be rebuilt from mech3ax (yet?).

Based on the WHATWG information on Windows-1252, which is itself a simple
single-byte encoding, it's trivial to write a decoder.